+++ CHANGELOG.md
@@ -29,6 +29,8 @@
 
 ### Added
 
+- [#3715](https://github.com/ChainSafe/forest/issues/3715): Implemented parallel HTTP downloads for snapshots with 5 concurrent connections by default (configurable via `FOREST_DOWNLOAD_CONNECTIONS`), bringing significant performance improvements for snapshot downloads (on par with a manual `aria2c -x5`).
+
 ### Changed
 
 ### Removed